Commit all

+++ admin/src/components/TableDataChildPitch/TableDataChildPitch.jsx
@@ -4,14 +4,23 @@ import { Link, useLocation, useParams } from "react-router-dom";
 import { useEffect, useState } from "react";
 import useFetch from "../../hooks/useFetch";
 import axios from "axios";
+import {
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogTitle,
+  DialogContentText,
+} from "@mui/material";
 
-
-const TableDataChildPitch = ({ columns }) => {
+const TableDataChildPitch = ({ columns, params }) => {
   const location = useLocation();
   const path = location.pathname.split("/")[1];
+  const [open, setOpen] = useState(false);
   const [list, setList] = useState([]);
   const { data } = useFetch(`/${path}`);
   const { pitchId } = useParams();
+  const [idDelete, setidDelete] = useState(null);
 
   useEffect(() => {
     setList(data);
@@ -19,9 +28,33 @@ const TableDataChildPitch = ({ columns }) => {
 
   const handleDelete = async (id) => {
     try {
-      await axios.delete(`/${path}/${id}/${pitchId}`);
-      setList(list.filter((item) => item._id !== id));
+      if (idDelete) {
+        const resp = await axios.delete(`/${path}/${id}/${pitchId}`);
+        if (resp) {
+          handleClose();
+          reFetch();
+        }
+      }
+      // setList(list.filter((item) => item._id !== id));
     } catch (err) {}
+  };
+
+  const reFetch = async () => {
+    try {
+      const res = await axios.get(path);
+      setList(res.data);
+    } catch (err) {
+      throw err;
+    }
+  };
+
+  const handleClickOpen = (idRow) => {
+    setOpen(true);
+    setidDelete(idRow);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
   };
 
   const actionColumn = [
@@ -40,7 +73,8 @@ const TableDataChildPitch = ({ columns }) => {
             </Link>
             <div
               className="deleteButton"
-              onClick={() => handleDelete(params.row._id)}
+              // onClick={() => handleClickOpen(params.row._id)}
+              onClick= {() => handleDelete(params.row._id)}
             >
               Delete
             </div>
@@ -71,6 +105,27 @@ const TableDataChildPitch = ({ columns }) => {
         checkboxSelection
         getRowId={(row) => row._id}
       />
+      <Dialog
+        open={open}
+        onClose={handleClose}
+        aria-labelledby="alert-dialog-title"
+        aria-describedby="alert-dialog-description"
+      >
+        <DialogTitle id="alert-dialog-title">
+          {"Delete Dialog when to use Delete!!!"}
+        </DialogTitle>
+        <DialogContent>
+          <DialogContentText id="alert-dialog-description">
+            Are you sure to delete this item?
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions>
+          <Button onClick={handleClose}>Cancel</Button>
+          <Button onClick={() => handleDelete()} autoFocus>
+            Delete
+          </Button>
+        </DialogActions>
+      </Dialog>
     </div>
   );
 };
